refactor: 메시지 미리보기 컴포넌트 분리
