+++ src/pages/match/message/components/topPreivew.module.css
@@ -1,0 +1,36 @@
+.matchs_container {
+  width: 100%;
+  height: 10rem;
+  padding: 1rem;
+  display: flex;
+  gap: 1rem;
+  background-color: black;
+}
+
+.matches_image {
+  opacity: 0;
+  animation: fadeInAndSlideUp 0.6s ease forwards;
+
+  width: 90px;
+  height: 100%;
+  background-size: cover;
+  background-repeat: no-repeat;
+  background-position: center;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.matches_image:hover {
+  border: 1px solid rgb(175, 175, 175);
+}
+
+@keyframes fadeInAndSlideUp {
+  0% {
+    opacity: 0;
+    transform: translateY(20px);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
